Chrome blocks image loading because of CORS security
According to Shiffman in
https://github.com/processing/p5.js/issues/561, one should be able to
call createImg() in p5 and then image() to draw it on the canvas.  If
so, that could work around Chrome blocking access to the image because
of cross-origin security.

It does work to use createImg() which returns a p5.Element instead of
loadImage() which returns a p5.Image.  Why I don’t know but now we can
preload images for all the small turrets and bullets and we do not have
any cross-origin security blocks when running without a server.